requirements.txt
chardet==4.0.0
coverage==4.5.3
Django==2.2.13
django-ckeditor==6.0.0
django-crispy-forms==1.13.0
django-debug-toolbar==1.11
django-filter==2.1.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-newsletter==0.9
django-taggit==1.1.0
djangorestframework==3.9.2
ldif3==3.1.1
model-mommy==1.6.0
Pillow==6.2.1
python-card-me==0.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
surlex==0.2.0
unicodecsv==0.14.1
